Clean up apt-get artifacts to reduce image sizes
Updates the Dockerfile generation scripts to include cleanup for apt-get's extraneous artifacts

This removes 30 MB from the base family and 20 MB from the included family. In production, it will reduce the included family by 50 MB because it will inherit 30 MB less from the base family + 20 MB reduction in the image itself.

There's a lot of bloat in the browsers family of images. They have several RUN commands that call apt-get and fail to clean up, so there's a few hundred MB of bloat from those in the final included family. It's hard to fix at this point because we don't have generators for them and we don't want to modify the image tags already published.