feat(CONTRIBUTING.md): simplify tagging instructions
Previously in order to create a tag, the actor had to download the image,
tag it locally, and then upload the new image.

Instead, introduce a way to create psuedo-tags for simple images that
obviates the requirement to download images.

Here we create a manifest containing a reference to the currently tagged image(s).