Update factory install scripts to check for supported arch (#875)
Prior to this change the install scripts explicitly check for arm64
arch and short-circuit on the browser install scripts since there
are no linux/arm64 binaries available. However, this assumes that
the only other target arch is x64, preventing the image from being
successfully build on other platforms. For example, the node install
script supports a variety of architectures, including ppc64le, s390x, etc.

Update the install scripts to check for supported platforms instead,
this is mostly x64, with the exception being webkit which is available
on both x64 and arm64. For all other values, short-circuit as before.

This means that the factory image can now be built for additional
platforms using docker buildx for example without running into errors
due to missing browser binaries.